REF: move masked_shared back into test_integer
